staticfiles/requirements.txt
django==6.0.3
djangorestframework
django-cors-headers
gunicorn
whitenoise
Pillow
